refactor: start putting some order in the @shared directory
