+++ src/routes/auth/change-email/change.ts
@@ -1,24 +1,20 @@
 import { Request, Response } from 'express'
-import {
-  asyncWrapper,
-  rotateTicket,
-  HasuraAccountData,
-  HasuraUpdateAccountData
-} from '@shared/helpers'
+import { asyncWrapper, rotateTicket } from '@shared/helpers'
 import { changeEmailByTicket, getNewEmailByTicket } from '@shared/queries'
 
 import Boom from '@hapi/boom'
 import { request } from '@shared/request'
-import { verifySchema } from '@shared/schema'
+import { verifySchema } from '@shared/validation'
+import { QueryAccountData, UpdateAccountData } from '@shared/types'
 
 async function changeEmail({ body }: Request, res: Response): Promise<unknown> {
   const { ticket } = await verifySchema.validateAsync(body)
 
   const {
     auth_accounts: [{ new_email }]
-  } = await request<HasuraAccountData>(getNewEmailByTicket, { ticket })
+  } = await request<QueryAccountData>(getNewEmailByTicket, { ticket })
 
-  const hasuraData = await request<HasuraUpdateAccountData>(changeEmailByTicket, {
+  const hasuraData = await request<UpdateAccountData>(changeEmailByTicket, {
     ticket,
     new_email,
     now: new Date()
