+++ src/routes/auth/providers/apple.ts
@@ -3,7 +3,7 @@ import { Strategy, Profile } from '@nicokaiser/passport-apple'
 import Boom from '@hapi/boom'
 import { PROVIDERS } from '@shared/config'
 import { initProvider } from './utils'
-import { UserData } from '@shared/helpers'
+import { UserData } from '@shared/types'
 
 const transformProfile = ({ id, name, email, photos }: Profile): UserData => ({
   id,
