+++ src/routes/auth/token/refresh.ts
@@ -1,12 +1,13 @@
-import { AccountData, asyncWrapper, createHasuraJwt, newRefreshExpiry } from '@shared/helpers'
+import { asyncWrapper } from '@shared/helpers'
 import { Request, Response } from 'express'
 import { selectRefreshToken, updateRefreshToken } from '@shared/queries'
 
 import Boom from '@hapi/boom'
 import { COOKIE_SECRET } from '@shared/config'
-import { newJwtExpiry } from '@shared/jwt'
+import { newJwtExpiry, newRefreshExpiry, createHasuraJwt } from '@shared/jwt'
 import { request } from '@shared/request'
 import { v4 as uuidv4 } from 'uuid'
+import { AccountData } from '@shared/types'
 
 interface HasuraData {
   auth_refresh_tokens: { account: AccountData }[]
